tests/wpt/web-platform-tests/html/canvas/element/fill-and-stroke-styles/2d.fillStyle.CSSRGB.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<title>Canvas test: 2d.fillStyle.CSSRGB</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.fillStyle.CSSRGB</h1>
<p class="desc">CSSRGB works as color input</p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:<p><img src="/images/green-100x50.png" class="output expected" id="expected" alt="">
<ul id="d"></ul>
<script>
var t = async_test("CSSRGB works as color input");
_addTest(function(canvas, ctx) {

ctx.fillStyle = new CSSRGB(1, 0, 1);
_assertSame(ctx.fillStyle, '#ff00ff', "ctx.fillStyle", "'#ff00ff'");
ctx.fillRect(0, 0, 100, 50);
_assertPixel(canvas, 50,25, 255,0,255,255, "50,25", "255,0,255,255");

const color = new CSSRGB(0, CSS.percent(50), 0);
ctx.fillStyle = color;
_assertSame(ctx.fillStyle, '#008000', "ctx.fillStyle", "'#008000'");
ctx.fillRect(0, 0, 100, 50);
_assertPixel(canvas, 50,25, 0,128,0,255, "50,25", "0,128,0,255");
color.g = 0;
ctx.fillStyle = color;
_assertSame(ctx.fillStyle, '#000000', "ctx.fillStyle", "'#000000'");
ctx.fillRect(0, 0, 100, 50);
_assertPixel(canvas, 50,25, 0,0,0,255, "50,25", "0,0,0,255");

color.alpha = 0;
ctx.fillStyle = color;
_assertSame(ctx.fillStyle, 'rgba(0, 0, 0, 0)', "ctx.fillStyle", "'rgba(0, 0, 0, 0)'");
ctx.reset();
color.alpha = 0.5;
ctx.fillStyle = color;
ctx.fillRect(0, 0, 100, 50);
_assertPixel(canvas, 50,25, 0,0,0,128, "50,25", "0,0,0,128");

ctx.fillStyle = new CSSHSL(CSS.deg(0), 1, 1).toRGB();
_assertSame(ctx.fillStyle, '#ffffff', "ctx.fillStyle", "'#ffffff'");
ctx.fillRect(0, 0, 100, 50);
_assertPixel(canvas, 50,25, 255,255,255,255, "50,25", "255,255,255,255");

color.alpha = 1;
color.g = 1;
ctx.fillStyle = color;
ctx.fillRect(0, 0, 100, 50);


});
</script>
